makerfair-rome-1

+++ pantallas.py
@@ -2,10 +2,7 @@ import lvgl as lv
 import guiObj1
 import guiObj2
 import guiObj3
-import guiObj4
-import guiObj5
-import guiObj6
-#import guiObj7
+import guiObj0
 import teclado
 
 #####################################
@@ -20,11 +17,7 @@ pantalla_sel = None
 Gal_pantallas = [{"Name":"Calculadora","function":guiObj1.guiObj1}
                  ,{"Name":"Editor","function":guiObj3.guiObj3 }
                  ,{"Name":"Graficas","function":guiObj2.guiObj2}
-                 ,{"Name":"Domotica","function":guiObj4.guiObj4}
-                 ,{"Name":"CibusTab","function":guiObj5.guiObj5}
-                 ,{"Name":"Peso","function":guiObj6.guiObj6}
-#                  ,{"Name":"CLI","function":guiObj7.guiObj7}
-#                  ,{"Name":"CLI","function":guiObj7.guiObj7}
+                 ,{"Name":"Demo","function":guiObj0.guiObj0}
                  ]
 
 
@@ -36,6 +29,8 @@ def pantalla_sel_event_handler(evt):
     if code == lv.EVENT.VALUE_CHANGED :
         id = obj.get_selected_btn()
         txt = obj.get_btn_text(id)
+        if(txt == "Void"):
+            return
         pantalla_sel.delete()
         miTeclado = teclado.teclado()
         miTeclado.graphCursor = None
@@ -56,17 +51,22 @@ def select():
         pantalla_sel.delete()
         pantalla_sel=None
         return
-    
+
     iColumnas=1
     i=0
     btnm_map =[]
     for pantalla in Gal_pantallas:
         btnm_map.append(pantalla['Name'])
+        print( pantalla['Name'] )
         i=i+1
         if i > iColumnas:
             btnm_map.append("\n")
             i=0
-    btnm_map[len(btnm_map)-1]=""    
+    if (len(Gal_pantallas)%2==1):   # screen number odd
+        print("3 screens")
+        btnm_map.append("Void")
+    else:                           # screen number even
+        btnm_map[len(btnm_map)-1]=""
     
     pantalla_sel = lv.btnmatrix(lv.scr_act())
     pantalla_sel.set_size(300, 200)
